+++ gazon-service/src/test/java/com/gitlab/service/PaymentServiceTest.java
@@ -5,17 +5,13 @@ import com.gitlab.dto.OrderDto;
 import com.gitlab.dto.PaymentDto;
 import com.gitlab.enums.OrderStatus;
 import com.gitlab.enums.PaymentStatus;
-import com.gitlab.mapper.PaymentMapper;
 import com.gitlab.model.BankCard;
-import com.gitlab.model.Payment;
 import com.gitlab.model.User;
-import com.gitlab.repository.PaymentRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
@@ -31,13 +27,7 @@ import static org.mockito.Mockito.*;
 class PaymentServiceTest {
 
     @Mock
-    private PaymentRepository paymentRepository;
-
-    @Mock
     private UserService userService;
-
-    @Mock
-    private PaymentMapper paymentMapper;
 
     @Mock
     private PaymentClient paymentClient;
@@ -50,10 +40,12 @@ class PaymentServiceTest {
 
     @Test
     void should_find_all_payments() {
-        List<Payment> expectedResult = generatePayments();
-        when(paymentRepository.findAll((Sort) any())).thenReturn(generatePayments());
+        List<PaymentDto> expectedResult = generatePaymentsDtos();
+        ResponseEntity<List<PaymentDto>> responseEntity = new ResponseEntity<>(expectedResult, HttpStatus.OK);
 
-        List<Payment> actualResult = paymentService.findAll();
+        when(paymentClient.getPaymentsPage(null, null)).thenReturn(responseEntity);
+
+        List<PaymentDto> actualResult = paymentService.findAllDto();
 
         assertEquals(expectedResult, actualResult);
     }
@@ -61,20 +53,23 @@ class PaymentServiceTest {
     @Test
     void should_find_payment_by_id() {
         long id = 1L;
-        Payment expectedResult = generatePayment();
-        when(paymentRepository.findById(id)).thenReturn(Optional.of(expectedResult));
+        PaymentDto expectedResult = generatePaymentDto();
+        ResponseEntity<Optional<PaymentDto>> responseEntity = new ResponseEntity<>(Optional.of(expectedResult), HttpStatus.OK);
 
-        Optional<Payment> actualResult = paymentService.findById(id);
+        when(paymentClient.getPaymentById(id)).thenReturn(responseEntity);
+
+        Optional<PaymentDto> actualResult = paymentService.findPaymentByIdDto(id);
 
         assertEquals(expectedResult, actualResult.orElse(null));
     }
 
     @Test
     void should_save_payment() {
-        Payment expectedResult = generatePayment();
-        when(paymentRepository.save(expectedResult)).thenReturn(expectedResult);
+        PaymentDto expectedResult = generatePaymentDto();
+        expectedResult.setPaymentStatus(PaymentStatus.NOT_PAID);
+        when(paymentClient.makePayment(expectedResult)).thenReturn(new ResponseEntity<>(expectedResult, HttpStatus.OK));
 
-        Payment actualResult = paymentService.save(expectedResult);
+        PaymentDto actualResult = paymentService.saveDto(expectedResult);
 
         assertEquals(expectedResult, actualResult);
     }
@@ -84,20 +79,13 @@ class PaymentServiceTest {
         User user = new User();
         user.setId(123L);
         Set<BankCard> userCards = new HashSet<>();
-        userCards.add(generatePayment().getBankCard());
+        userCards.add(new BankCard());
         user.setBankCardsSet(userCards);
 
-        PaymentDto paymentDto = new PaymentDto();
+        PaymentDto paymentDto = generatePaymentDto();
         paymentDto.setPaymentStatus(PaymentStatus.NOT_PAID);
 
-        Payment payment = new Payment();
-
-        Payment savedPayment = new Payment();
-        PaymentDto savedPaymentDto = new PaymentDto();
-        savedPaymentDto.setPaymentStatus(PaymentStatus.PAID);
-        savedPaymentDto.setOrderId(1L);
-
-        PaymentDto paymentDtoResponse = new PaymentDto();
+        PaymentDto paymentDtoResponse = generatePaymentDto();
         paymentDtoResponse.setPaymentStatus(PaymentStatus.PAID);
         paymentDtoResponse.setOrderId(1L);
 
@@ -106,13 +94,6 @@ class PaymentServiceTest {
         OrderDto orderDto = new OrderDto();
         orderDto.setOrderStatus(OrderStatus.NOT_PAID);
 
-        when(userService.getAuthenticatedUser()).thenReturn(user);
-        when(paymentMapper.toEntity(paymentDto)).thenReturn(payment);
-        when(paymentRepository.save(payment)).thenReturn(savedPayment);
-        Payment savedDtoResponseEntity = new Payment();
-        savedDtoResponseEntity.setId(1L);
-        when(paymentRepository.save(paymentMapper.toEntity(paymentDtoResponse))).thenReturn(savedDtoResponseEntity);
-        when(paymentMapper.toDto(savedPayment)).thenReturn(savedPaymentDto);
         when(paymentClient.makePayment(any(PaymentDto.class))).thenReturn(paymentDtoResponseEntity);
         when(orderService.findByIdDto(1L)).thenReturn(Optional.of(orderDto));
 
@@ -123,38 +104,20 @@ class PaymentServiceTest {
         assertEquals(OrderStatus.PAID, orderDto.getOrderStatus());
     }
 
-    @Test
-    void should_delete_payment() {
-        long id = 1L;
-        when(paymentRepository.findById(id)).thenReturn(Optional.of(generatePayment()));
 
-        paymentService.delete(id);
 
-        verify(paymentRepository).deleteById(id);
-    }
-
-    @Test
-    void should_not_delete_payment_when_entity_not_found() {
-        long id = 1L;
-        when(paymentRepository.findById(id)).thenReturn(Optional.empty());
-
-        paymentService.delete(id);
-
-        verify(paymentRepository, never()).deleteById(anyLong());
-    }
-
-    private List<Payment> generatePayments() {
+    private List<PaymentDto> generatePaymentsDtos() {
         return List.of(
-                new Payment(1L),
-                new Payment(2L),
-                new Payment(3L),
-                new Payment(4L),
-                new Payment(5L),
-                new Payment(6L)
+                new PaymentDto(),
+                new PaymentDto(),
+                new PaymentDto(),
+                new PaymentDto(),
+                new PaymentDto(),
+                new PaymentDto()
         );
     }
 
-    private Payment generatePayment() {
-        return new Payment(1L);
+    private PaymentDto generatePaymentDto() {
+        return new PaymentDto();
     }
 }